library drawer working
